+++ clang/test/Preprocessor/has_c_attribute.c
@@ -33,12 +33,45 @@ C2x(__gnu__::warn_unused_result)
 // CHECK: gnu::__warn_unused_result__: 201904L
 C2x(gnu::__warn_unused_result__)
 
-// We do somewhat support the __clang__ vendor namespace, but it is a
-// predefined macro and thus we encourage users to use _Clang instead.
-// Because of this, we do not support __has_c_attribute for that
-// vendor namespace.
-//
-// Note, we can't use C2x here because it will expand __clang__ to 1
-// too early.
-// CHECK: 1::fallthrough: 0
-__clang__::fallthrough: __has_c_attribute(__clang__::fallthrough)
+// Test that macro expansion of the builtin argument works.
+#define C clang
+#define L likely
+#define CL clang::likely
+#define N nodiscard
+
+#if __has_c_attribute(N)
+int has_nodiscard;
+#endif
+// CHECK: int has_nodiscard;
+
+#if __has_c_attribute(C::L)
+int has_clang_likely_1;
+#endif
+// CHECK: int has_clang_likely_1;
+
+#if __has_c_attribute(clang::L)
+int has_clang_likely_2;
+#endif
+// CHECK: int has_clang_likely_2;
+
+#if __has_c_attribute(C::likely)
+int has_clang_likely_3;
+#endif
+// CHECK: int has_clang_likely_3;
+
+#if __has_c_attribute(CL)
+int has_clang_likely_4;
+#endif
+// CHECK: int has_clang_likely_4;
+
+#define FUNCLIKE1(x) clang::x
+#if __has_c_attribute(FUNCLIKE1(likely))
+int funclike_1;
+#endif
+// CHECK: int funclike_1;
+
+#define FUNCLIKE2(x) _Clang::x
+#if __has_c_attribute(FUNCLIKE2(likely))
+int funclike_2;
+#endif
+// CHECK: int funclike_2;